% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaLike.R
\name{BetaLike}
\alias{BetaLike}
\title{Beta Loglikelihood}
\usage{
BetaLike(X, a1, b1, minDet = 1e-09)
}
\arguments{
\item{X}{data vector.}

\item{a1}{shape parameter 1 of the Beta distribution.}

\item{b1}{shape parameter 2 of the Beta distribution.}

\item{minDet}{value to replace any 0s or (1-minDet) for 1s, default minDet=1e-9.}
}
\value{
A statistic.
}
\description{
This function calculates the log likelihood of a given vector given the alpha and beta shape parameters of the Beta distribution. It uses the Stirling's approximation of the beta function.
}
\examples{
X=c(0.2, 0.4, 0.3, 0.6, 0.2, 0.1, 0.4)
BetaLike(X, a1=1, b1=1)

}
