% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaLikeRat.R
\name{BetaLikeRat}
\alias{BetaLikeRat}
\title{Beta Loglikelihood ratio}
\usage{
BetaLikeRat(X, a1, b1, a2, b2, minDet = 1e-09)
}
\arguments{
\item{X}{data vector.}

\item{a1}{shape parameter 1 of first candidate Beta distribution.}

\item{b1}{shape parameter 2 of first candidate Beta distribution.}

\item{a2}{shape parameter 1 of second candidate Beta distribution.}

\item{b2}{shape parameter 2 of second candidate Beta distribution.}

\item{minDet}{value to replace any 0s or (1-minDet) for 1s, default 1e-9.}
}
\value{
A statistic with large values indicating more support for the first candidate Beta distribution.
}
\description{
This function assesses the goodness of fit of a given vector for two candidate Beta distributions. Calculates Beta likelihood ratio for two Beta distributions with two sets of parameters and a data vector.
}
\examples{
X=c(0.2, 0.4, 0.3, 0.6, 0.2, 0.1, 0.4)
BetaLikeRat(X, a1=1, b1=1, a2=10, b2=5)
}
