% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaParEst.R
\name{BetaParEst}
\alias{BetaParEst}
\title{Beta distribution}
\usage{
BetaParEst(m, v)
}
\arguments{
\item{m}{data mean.}

\item{v}{data variance.}
}
\value{
A list containing alpha and beta parameters of the Beta distribution.
}
\description{
This function calculates the alpha and beta shape parameters of the Beta distribution via the method of moments.
}
\examples{
BetaParEst(m=0.5, v=0.2)
}
