% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Disp.BarrierSym.R
\docType{data}
\name{Disp.BarrierSym}
\alias{Disp.BarrierSym}
\title{Symmetrical dispersal with gene flow barrier dispersal matrix}
\description{
Symmetrical dispersal with gene flow barrier dispersal matrix
}
