% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Disp.BarrierUni.R
\docType{data}
\name{Disp.BarrierUni}
\alias{Disp.BarrierUni}
\title{Unidirectional dispersal with gene flow barrier dispersal matrix}
\description{
Unidirectional dispersal with gene flow barrier dispersal matrix
}
