% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dispsim.R
\name{Disp.sim}
\alias{Disp.sim}
\title{Dispersal Generated Allele Frequencies}
\usage{
Disp.sim(
  Dispersal = "new",
  Dispersion = NA,
  N = 1000,
  PopLoc = 10,
  YearSteps = 1000,
  MinAge = 5,
  MaxAge = 20,
  loci = 1000,
  Mutation = TRUE,
  Selection = FALSE,
  selstr = 0,
  selrand = FALSE,
  Stoch = TRUE,
  mut = 1e-06
)
}
\arguments{
\item{Dispersal}{Use the specified Dispersion matrix, or alternatively if "Random" calls Dispers.Rand function to create random dispersal matrix at each year step.}

\item{Dispersion}{Matrix of dispersion probabilities.}

\item{N}{Effective population size}

\item{PopLoc}{Number of populations to model}

\item{YearSteps}{Number of year steps to simulate}

\item{MinAge}{Minimum age of breeding cohort}

\item{MaxAge}{Maximum age of oldest cohort, i.e. maximum lifespan}

\item{loci}{Number of loci to simulate}

\item{Mutation}{Whether to include a mutation parameter in the simulation}

\item{Selection}{Whether to include a selection parameter in the simulation}

\item{selstr}{Selection strength parameter [-1,1] where each cohort selection~unif(1, -selstr, selstr)}

\item{selrand}{Whether selection is random for each loci~unif(loci, -selstr, selstr)}

\item{Stoch}{Whether to model the dispersion stochastically where each generation parents are chosen at random}

\item{mut}{mutation rate}
}
\value{
A list containing allele frequencies for each age class and population
}
\description{
This function the main stochastic genetic population algorithm that generates the allele frequencies based on dispersal matrix.
}
\examples{
Disp.sim(Dispersal="Random", Dispersion=NA, N=1e3, PopLoc=10, YearSteps=1000, MinAge=5, MaxAge=20,loci=1000, Mutation=TRUE, Selection=FALSE, selstr=0, selrand=FALSE, SStoch=TRUE, mut=1e-6)
}
