% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispStationaritySim.R
\name{Disp.Stationarity.sim}
\alias{Disp.Stationarity.sim}
\title{Time to Stationarity}
\usage{
Disp.Stationarity.sim(
  Dispersal = "new",
  Dispersion = NA,
  N = 1000,
  PopLoc = 10,
  MinAge = 5,
  MaxAge = 20,
  loci = 1000,
  Mutation = TRUE,
  Selection = FALSE,
  selstr = 1,
  Stoch = TRUE,
  mut = 1e-06,
  YStart = 1,
  YInterval = 10,
  YFinal = 1000
)
}
\arguments{
\item{Dispersal}{Use the specified Dispersion matrix, or alternatively if "Random" calls Dispers.Rand function to create random dispersal matrix at each year step.}

\item{Dispersion}{Matrix of dispersion probabilities.}

\item{N}{Effective population size.}

\item{PopLoc}{Number of populations to model}

\item{MinAge}{Minimum age of breeding cohort}

\item{MaxAge}{Maximum age of oldest cohort, i.e. maximum lifespan}

\item{loci}{Number of loci to simulate}

\item{Mutation}{Whether to include a mutation parameter in the simulation}

\item{Selection}{Whether to include a selection parameter in the simulation}

\item{selstr}{Selection strength parameter [-1,1]}

\item{Stoch}{Whether to model the dispersion stochastically where each generation parents are chosen at random}

\item{mut}{mutation rate}

\item{YStart}{Start year for recording stationarity variables}

\item{YInterval}{Interval years between recording stationarity variables}

\item{YFinal}{Stop simulation after these many year steps}
}
\value{
A data frame of statistics between time intervals

yr1, yr2: year intervals

corr: correlation between the population Fst matrix in yr1 and yr2

meanFst: mean population Fst in yr2

varFst: variance of population Fst in yr2
}
\description{
Function to test for number of generations to stationarity by calculating correlation, mean and variance for allele frequencies between YInterval generations:
}
\examples{
Disp.Stationarity.sim(Dispersal="Random", Dispersion=NA, N=1e3, PopLoc=10, MinAge=5, MaxAge=20,loci=1000, Mutation=TRUE, Selection=FALSE, selstr=1, Stoch=TRUE, mut=1e-6, YStart=1, YInterval=10, YFinal=1e3)
}
