% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispStationSlope.R
\name{Disp.StationSlope}
\alias{Disp.StationSlope}
\title{Time to Stationarity Slope}
\usage{
Disp.StationSlope(
  Dispersal = "new",
  Dispersion = NA,
  N = 1000,
  PopLoc = 10,
  MinAge = 5,
  MaxAge = 20,
  loci = 1000,
  Mutation = TRUE,
  Selection = FALSE,
  selstr = 0,
  Stoch = TRUE,
  mut = 1e-06,
  YStart = 1,
  YInterval = 100,
  YFinal = 1000,
  tolerance = 6
)
}
\arguments{
\item{Dispersal}{Use the specified Dispersion matrix, or alternatively if "Random" calls Dispers.Rand function to create random dispersal matrix at each year step.}

\item{Dispersion}{Specify the dispersal matrix}

\item{N}{Effective population size}

\item{PopLoc}{Number of populations}

\item{MinAge}{Minimum age for spawning cohort}

\item{MaxAge}{Maximum lifespan}

\item{loci}{Number of loci to simulate}

\item{Mutation}{Whether to incorporate mutation}

\item{Selection}{Whether to incorporate selection}

\item{selstr}{Selection strength}

\item{Stoch}{Whether to include stochasticity}

\item{mut}{Mutation rate}

\item{YStart}{Start year for recording stationarity variables}

\item{YInterval}{Interval years between testing slope}

\item{YFinal}{Stop simulation after these many generations}

\item{tolerance}{The number of significant digits to use for calculating significance of slope test}
}
\value{
A data frame containing two year values

SlopeMin year when first occurrence of significant 0 slope for mean and variance of Fst.

SlopeMax year when last occurrence of significant 0 slope for mean and variance of Fst.
}
\description{
Function calculates the significance of the slope for the mean pairwise Fst for a given dispersal model and time frame.
}
\examples{
Disp.StationSlope(Dispersal="Random", Dispersion=NA, N=1e3, PopLoc=10, MinAge=5, MaxAge=20, loci=1000, Mutation=TRUE, Selection=FALSE, selstr=0, Stoch=TRUE, mut=1e-6, YStart=1, YInterval=100, YFinal=1e3, tolerance=6)
}
