% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Disp.tst.R
\docType{data}
\name{Disp.tst}
\alias{Disp.tst}
\title{Allele frequencies of the simulated dispersal examples
\itemize{
  \item Dispersal. Dispersion type
  \item Population. Population
  \item Age. Age of cohort
  \item Loci. Name of loci (numbered)
  \item Dispersion.colors. Dispersal type corresponding colors
  \item DispYears. Number of years to stationarity
}}
\format{
A list.
}
\usage{
data(Disp.tst)
}
\description{
Allele frequencies of the simulated dispersal examples
\itemize{
  \item Dispersal. Dispersion type
  \item Population. Population
  \item Age. Age of cohort
  \item Loci. Name of loci (numbered)
  \item Dispersion.colors. Dispersal type corresponding colors
  \item DispYears. Number of years to stationarity
}
}
\references{
R base package.
}
\keyword{datasets}
