% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersRand.R
\name{Dispers.Rand}
\alias{Dispers.Rand}
\title{Random Dispersal}
\usage{
Dispers.Rand(Pops = 10)
}
\arguments{
\item{Pops}{Number of populations. Default Pops=10.}
}
\value{
A Pops*Pops matrix of dispersal parameters.
}
\description{
This function calculates the rescaled dispersal matrix with rows summing to 1
}
\examples{
Dispers.Rand(Pops=10)
}
