% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datat-Dispersal.models.R
\docType{data}
\name{Dispersal.models}
\alias{Dispersal.models}
\title{Parameters for all the example dispersal models
\itemize{
  \item Dispersion. Dispersion type
  \item DispModel. Specifies the corresponding dispersal matrix
  \item DispMinAge. Minimum breeding age
  \item DispMaxAge. Age of death
  \item Dispersion.colors. Dispersal type corresponding colors
  \item DispYears. Number of years to stationarity
}}
\format{
A data frame.
}
\usage{
data(Dispersal.models)
}
\description{
Parameters for all the example dispersal models
\itemize{
  \item Dispersion. Dispersion type
  \item DispModel. Specifies the corresponding dispersal matrix
  \item DispMinAge. Minimum breeding age
  \item DispMaxAge. Age of death
  \item Dispersion.colors. Dispersal type corresponding colors
  \item DispYears. Number of years to stationarity
}
}
\references{
R base package.
}
\keyword{datasets}
