% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Dispersal.StationT.R
\docType{data}
\name{Dispersal.StationT}
\alias{Dispersal.StationT}
\title{Time to stationarity for all the dispersal examples
\itemize{
  \item DispModel. Dispersion type
  \item SlopeMin. First year when mean Fst and variance Fst did not change
  \item SlopeMax. Last year when mean Fst and variance Fst no longer changed afterwards
}}
\format{
A data frame.
}
\usage{
data(Dispersal.StationT)
}
\description{
Time to stationarity for all the dispersal examples
\itemize{
  \item DispModel. Dispersion type
  \item SlopeMin. First year when mean Fst and variance Fst did not change
  \item SlopeMax. Last year when mean Fst and variance Fst no longer changed afterwards
}
}
\references{
R base package.
}
\keyword{datasets}
