% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-distances.R
\docType{data}
\name{distances}
\alias{distances}
\title{Unit distance matrix used for the dispersal models}
\format{
A distance matrix.
}
\usage{
data(distances)
}
\description{
Unit distance matrix used for the dispersal models
}
\references{
R base package.
}
\keyword{datasets}
