% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fstmat.R
\name{Fst.mat}
\alias{Fst.mat}
\title{Pairwise Gst matrix}
\usage{
Fst.mat(Pop)
}
\arguments{
\item{Pop}{Population allele frequency object as generated from Disp.sim function}
}
\value{
A distance matrix.
}
\description{
This function calculates the pairwise G"st matrix between K populations averaging over age classes (see Nei and Chesser, 1983)
}
\examples{
Pop=Disp.sim(Dispersal="Random", Dispersion=NA, N=1e3, PopLoc=10, YearSteps=1000, MinAge=5, MaxAge=20,loci=1000, Mutation=TRUE, Selection=FALSE, selstr=1, Stoch=TRUE, mut=1e-6)
Fst.mat(Pop)

}
