% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWEchisq.R
\name{HWE.chisq}
\alias{HWE.chisq}
\title{HWE test}
\usage{
HWE.chisq(Gen)
}
\arguments{
\item{Gen}{Genotype matrix of n individual by l loci}
}
\value{
A vector.
}
\description{
This function calculate HWE Chisquare test p-value for genotype matrix:
}
\examples{
Gen=matrix(rbinom(100, 2, 0.5), ncol=20)
HWE.chisq(Gen)
}
