% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iserr.R
\name{is.err}
\alias{is.err}
\title{Error handling}
\usage{
is.err(code)
}
\arguments{
\item{code}{code to be tested to handle errors}
}
\value{
error message
}
\description{
Used to handle errors internally generated by functions without breaking the function
}
\examples{
x=matrix(runif(10,0,1),ncol=2)
is.err(x[,4])
}
