% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LociFst.R
\name{LociFst}
\alias{LociFst}
\title{Pairwise Fst for loci}
\usage{
LociFst(Pop1, Pop2)
}
\arguments{
\item{Pop1}{Population 1 allele frequency matrix}

\item{Pop2}{Population 2 allele frequency matrix}
}
\value{
A vector.
}
\description{
This function calculates the pairwsie Fst for each loci between two populations allele frequency matrices.
}
\examples{
Pop1=matrix(rbeta(10,1,1))
Pop2=matrix(rbeta(10,1,10))
LociFst(Pop1, Pop2)

}
