% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManhattanPlots.R
\name{ManhattanPlots}
\alias{ManhattanPlots}
\title{Manhattan Plot}
\usage{
ManhattanPlots(Pop, i = 1, j = 2, AllAg = FALSE, Age = 1)
}
\arguments{
\item{Pop}{Object created by Disp.sim function}

\item{i}{which 1st populations to compare}

\item{j}{which 2nd populations to compare}

\item{AllAg}{Whether to calculate allele frequencies across all age groups}

\item{Age}{which age to use for single age allele frequencies}
}
\value{
A vector.
}
\description{
This function creates a vector of pairwise Fst calculated for each loci between two populations i and j from POP object
}
\examples{
Pop=Disp.sim(Dispersal="Random", Dispersion=NA, N=1e3, PopLoc=10, YearSteps=1000, MinAge=1, MaxAge=1,loci=100, Mutation=TRUE, Selection=FALSE, selstr=1, Stoch=TRUE, mut=1e-6)
ManhattanPlots(Pop, i=1, j=2, AllAg=FALSE, Age=1)

}
