% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-MLE.Xval.R
\docType{data}
\name{MLE.Xval}
\alias{MLE.Xval}
\title{MLE cross validation results for the example dispersal}
\format{
A data frame.
}
\usage{
data(MLE.Xval)
}
\description{
MLE cross validation results for the example dispersal
}
\references{
R base package.
}
\keyword{datasets}
