% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Vector Normalization}
\usage{
normalize(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
A vector.
}
\description{
This function normalizes a set of values to range (0:1). It is used in the RankPerc function for heatmap graphing.
}
\examples{
x=runif(100,min=0, max=1)
normalize(x)
}
