% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Par.table.R
\docType{data}
\name{Par.table}
\alias{Par.table}
\title{Maximum likelihood estimates for the alpha and beta parameters of Beta distribution for each dispersal example
\itemize{
  \item PopStructure. Dispersion type
  \item MLEa. MLE estimate of shape parameter alpha
  \item MLEb. MLE estimate of shape parameter beta
}}
\format{
A data frame.
}
\usage{
data(Par.table)
}
\description{
Maximum likelihood estimates for the alpha and beta parameters of Beta distribution for each dispersal example
\itemize{
  \item PopStructure. Dispersion type
  \item MLEa. MLE estimate of shape parameter alpha
  \item MLEb. MLE estimate of shape parameter beta
}
}
\references{
R base package.
}
\keyword{datasets}
