% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Params.loci.R
\docType{data}
\name{Params.loci}
\alias{Params.loci}
\title{Results of the effect of the number of loci on the precision of Beta parameters:
\itemize{
  \item Loci. Number of loci
  \item alpha. MLE estimate of shape parameter alpha
  \item beta. MLE estimate of shape parameter beta
  \item meanFST. Mean Fst for given number of loci
  \item varFST. Variance Fst for given number of loci
}}
\format{
A data frame.
}
\usage{
data(Params.loci)
}
\description{
Results of the effect of the number of loci on the precision of Beta parameters:
\itemize{
  \item Loci. Number of loci
  \item alpha. MLE estimate of shape parameter alpha
  \item beta. MLE estimate of shape parameter beta
  \item meanFST. Mean Fst for given number of loci
  \item varFST. Variance Fst for given number of loci
}
}
\references{
R base package.
}
\keyword{datasets}
