% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Params.R
\docType{data}
\name{Params}
\alias{Params}
\title{Maximum likelihood estimates for the alpha and beta parameters of Beta distribution for each dispersal example
\itemize{
  \item PopStructure. Dispersion type
  \item Mean. Mean Fst for each dispersal type
  \item Var. Fst variance for each dispersal type
  \item MLEa. MLE estimate of shape parameter alpha
  \item MLEb. MLE estimate of shape parameter beta
}}
\format{
A data frame.
}
\usage{
data(Params)
}
\description{
Maximum likelihood estimates for the alpha and beta parameters of Beta distribution for each dispersal example
\itemize{
  \item PopStructure. Dispersion type
  \item Mean. Mean Fst for each dispersal type
  \item Var. Fst variance for each dispersal type
  \item MLEa. MLE estimate of shape parameter alpha
  \item MLEb. MLE estimate of shape parameter beta
}
}
\references{
R base package.
}
\keyword{datasets}
