% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankPerc.R
\name{RankPerc}
\alias{RankPerc}
\title{Rank Percentiles}
\usage{
RankPerc(x)
}
\arguments{
\item{x}{matrix of values}
}
\value{
A matrix.
}
\description{
This function rescales the dispersal values for for graphing scaling visualization of relative proportions.
}
\examples{
x=matrix(runif(100,min=0, max=1), ncol=20)
RankPerc(x)
}
