% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SelResults.R
\docType{data}
\name{SelResults}
\alias{SelResults}
\title{Results of the effect of cohort specific selection on dispersal cross validation
\itemize{
  \item ModelDispersal. Dispersal type
  \item SelStrength. Selection strength
  \item DispersalAllCohorts. Loglikelihood results for best fit dispersal model when sampling All Cohorts
  \item DispersalRandAllCohorts. Loglikelihood results for best fit dispersal model when sampling random Cohorts
  \item DispersalSelCohorts. Loglikelihood results for best fit dispersal model when sampling only the cohort under selection
  \item DispersalRandSingleCohort. Loglikelihood results for best fit dispersal model when sampling a single random cohort
}}
\format{
A data frame.
}
\usage{
data(SelResults)
}
\description{
Results of the effect of cohort specific selection on dispersal cross validation
\itemize{
  \item ModelDispersal. Dispersal type
  \item SelStrength. Selection strength
  \item DispersalAllCohorts. Loglikelihood results for best fit dispersal model when sampling All Cohorts
  \item DispersalRandAllCohorts. Loglikelihood results for best fit dispersal model when sampling random Cohorts
  \item DispersalSelCohorts. Loglikelihood results for best fit dispersal model when sampling only the cohort under selection
  \item DispersalRandSingleCohort. Loglikelihood results for best fit dispersal model when sampling a single random cohort
}
}
\references{
R base package.
}
\keyword{datasets}
