% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StGst.R
\name{stGst}
\alias{stGst}
\title{Meirmans and Hedricks Gst}
\usage{
stGst(Pop1, Pop2)
}
\arguments{
\item{Pop1}{Population 1 allele frequency vector}

\item{Pop2}{Population 2 allele frequency vector}
}
\value{
A value.
}
\description{
This function calculates the pairwise G"st value for two allele frequency vectors based on Meirmans and Hedricks 2011: "Assessing population structure: FST and related measures".
}
\examples{
Pop1=runif(100, min=0, max=1)
Pop2=runif(100, min=0, max=1)
stGst(Pop1, Pop2)
}
