% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Str.groups.R
\docType{data}
\name{Str.groups}
\alias{Str.groups}
\title{STRUCTURE results of population assignments for each of the dispersal examples.}
\format{
A List.
}
\usage{
data(Str.groups)
}
\description{
STRUCTURE results of population assignments for each of the dispersal examples.
}
\references{
R base package.
}
\keyword{datasets}
