% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Structure.matrices.R
\docType{data}
\name{Structure.matrices}
\alias{Structure.matrices}
\title{Results of STRUCTURE analysis for each of the dispersal models}
\format{
A data frame.
}
\usage{
data(Structure.matrices)
}
\description{
Results of STRUCTURE analysis for each of the dispersal models
}
\references{
R base package.
}
\keyword{datasets}
