% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimetoStationarity.R
\name{TimetoStationarity}
\alias{TimetoStationarity}
\title{Matrix Stationarity}
\usage{
TimetoStationarity(Disp.matrix = c(), delta = 1e-06, steps = 1e+06)
}
\arguments{
\item{Disp.matrix}{Dispersal matrix}

\item{delta}{Tolerance parameter to denote stationarity}

\item{steps}{maximum number of iterations}
}
\value{
A distance matrix.
}
\description{
This function calculates time to stationarity based on the x%*%x' approach
}
\examples{
Disp.matrix=matrix(data=runif(100,min=0,max=1), nrow=10)#' Fst.mat(Pop)
TimetoStationarity(Disp.matrix)
}
